requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
Django==5.1.7
djangorestframework==3.15.2
fonttools==4.56.0
geopandas==1.0.1
idna==3.10
kiwisolver==1.4.8
networkx==3.4.2
numpy==2.2.4
osmnx==2.0.1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pyogrio==0.10.0
pyparsing==3.2.1
pyproj==3.7.1
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
shapely==2.0.7
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
